Payment data to db (#154)
add payment data to database

+++ service/src/main/kotlin/fi/espoo/vekkuli/config/Paytrail.kt
@@ -32,7 +32,7 @@ data class PaytrailCallbackUrl(
 data class PaytrailPurchaseItem(
     val unitPrice: Int,
     val units: Int,
-    val vatPercentage: Int,
+    val vatPercentage: Double,
     val productCode: String,
     val description: String? = null,
     val category: String? = null,
